pyproject.toml
[project]
name = "telegram-deposit-bot"
version = "0.1.0"
description = "Telegram bot for UPI deposits with SMS-based verification"
requires-python = ">=3.11"
dependencies = [
    "python-telegram-bot>=21.0",
    "fastapi==0.109.0",
    "uvicorn==0.27.0",
    "asyncpg>=0.30.0",
    "redis==5.0.1",
    "python-dotenv==1.0.0",
    "apscheduler==3.10.4",
    "httpx>=0.25.2",
    "Pillow>=11.0.0",
]

[project.scripts]
deposit-bot = "bot.main:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["bot"]

[dependency-groups]
dev = [
    "pytest==7.4.4",
    "pytest-asyncio==0.23.3",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["."]
python_files = ["test_*.py"]
